libshviotqt: Refactor serialportsocket to a separate file
WASM doesn't support Qt::SerialPort.
